+++ ChangeLog.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- ...
+- function `io/file-normalize-utf`